+++ website/docs/CHANGELOG.md
@@ -9,8 +9,9 @@ sidebar_label: Changelog
 ### Breaking Changes
 
 - #698 Removes some methods and classes, namely:
-  - Combined `SchemaFactory::addControllerNamespace()` and `SchemaFactory::addTypeNamespace()` into one `SchemaFactory::addNamespace()`
-  - Removed `SchemaFactory::setGlobTTL()`, `FactoryContext::get*TTL()` and `RootTypeMapperFactoryContext::get*TTL()` as GraphQLite no longer uses TTLs to invalidate caches
+  - Deprecated `SchemaFactory::addControllerNamespace()` and `SchemaFactory::addTypeNamespace()` in favor of `SchemaFactory::addNamespace()`
+  - Deprecated `SchemaFactory::setGlobTTL()` in favor of `SchemaFactory::devMode()` and `SchemaFactory::prodMode()`
+  - Removed `FactoryContext::get*TTL()` and `RootTypeMapperFactoryContext::get*TTL()` as GraphQLite no longer uses TTLs to invalidate caches
   - Removed `StaticClassListTypeMapper` in favor of `ClassFinderTypeMapper` used with `StaticClassFinder`
   - Renamed `GlobTypeMapper` to `ClassFinderTypeMapper`
   - Renamed `SchemaFactory::setClassBoundCacheContractFactory()` to `SchemaFactory::setClassBoundCache()`,